rename test packages
