+++ kin-devplatform-sdk/src/test/java/kin/devplatform/BaseTestClass.java
