+++ kin-devplatform-sdk/src/test/java/kin/devplatform/CommonModifierFake.java
