+++ kin-devplatform-sdk/src/test/java/kin/devplatform/balance/presenter/BalancePresenterTest.java
